+++ CHANGELOG.md
@@ -4,13 +4,22 @@
 
 ### New features
 
+* [#4544](https://github.com/xmake-io/xmake/issues/4544): Support to wait process timeout for `xmake test`
+
+## v2.8.6
+
+### New features
+
 * Add `network.mode` policy
-* [#1433](https://github.com/xmake-io/xmake/issues/1433): Add `xmake pack` command to generate NSIS/zip/tar.gz/runself packages like cmake/cpack
+* [#1433](https://github.com/xmake-io/xmake/issues/1433): Add `xmake pack` command to generate NSIS/zip/tar.gz/rpm/srpm/runself packages like cmake/cpack
 * [#4435](https://github.com/xmake-io/xmake/issues/4435): Support batchsize for UnityBuild in Group Mode
+* [#4485](https://github.com/xmake-io/xmake/pull/4485): Support package.install_locally
+* Support NetBSD
 
 ### Changes
 
 * [#4484](https://github.com/xmake-io/xmake/pull/4484): Improve swig rule
+* Improve Haiku support
 
 ### Bugs fixed
 
@@ -1699,13 +1708,22 @@
 
 ### 新特性
 
+* [#4544](https://github.com/xmake-io/xmake/issues/4544): 改进 `xmake test`，支持等待进程超时
+
+## v2.8.6
+
+### 新特性
+
 * 添加 `network.mode` 策略
-* [#1433](https://github.com/xmake-io/xmake/issues/1433): 添加 `xmake pack` 命令去生成 NSIS/zip/tar.gz/runself 安装包
+* [#1433](https://github.com/xmake-io/xmake/issues/1433): 添加 `xmake pack` 命令去生成 NSIS/zip/tar.gz/rpm/srpm/runself 安装包
 * [#4435](https://github.com/xmake-io/xmake/issues/4435): 为 UnityBuild 的组模式增加 batchsize 支持
+* [#4485](https://github.com/xmake-io/xmake/pull/4485): 新增 package.install_locally 策略支持
+* 新增 NetBSD 支持
 
 ### Changes
 
 * [#4484](https://github.com/xmake-io/xmake/pull/4484): 改进 swig 规则
+* 改进 Haiku 支持
 
 ### Bugs 修复